requirements.txt
openai==0.27.0
cryptography==3.4.7
python-dotenv==0.19.0
unittest2==1.1.0
